requirements-macos.txt
numpy
Pillow
opencv-python
PyOpenGL
PyQt6
termcolor
websocket_client
flask
flask_cors
pyreadline3
openai
gTTS
google-cloud-texttospeech
matplotlib
PyYAML
torch
torchvision
